+++ examples/common/mod.rs
@@ -1,0 +1,2 @@
+pub const TEXT: &str = "Sphinx of black quartz, judge my vow!";
+pub const FONT_WIDTH: usize = 5;